requirements.txt
Django==1.9.5
dj-database-url==0.4.0
django-floppyforms==1.6.1
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==3.0
